% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comet_tags}
\alias{comet_tags}
\title{comet_wa tags}
\format{
A data frame with 122 rows and 7 variables:
\describe{
  \item{class}{Broad class of conservation practice}
  \item{practice}{USDA-NRCS Conservation Practice Standard name}
  \item{implementation}{The implementation scenario modeled for each conservation practice standard; implementation scenarios are described in more detail in the one-page summaries associated with each practice}
  \item{abbr}{Abbreviated implementation for plots.}
  \item{current_land_use}{Value extracted from planner_implementation to use as filtering tag.}
  \item{irrigation}{Value extracted from planner_implementation to use as filtering tag.}
  \item{nutrient_practice}{Value extracted from planner_implementation to use as filtering tag.}
  }
}
\usage{
comet_tags
}
\description{
Tags for filtering conservation practice implementations from the comet_wa dataset
}
\keyword{datasets}
