% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comet_wa}
\alias{comet_wa}
\title{A subset of the NRCS Comet-Planner dataset}
\format{
A data frame with 17280 rows and 14 variables. The Rda dataset tidies the data by pivoting longer.
\describe{
  \item{state}{Abbreviation for state}
  \item{county}{County name}
  \item{fipsint}{County fips code (unique integer identifier for all U.S. counties)}
  \item{mlra}{USDA Major Land Resource Area rectified to county boundaries}
  \item{lrr}{USDA Land Resource Region rectified to county boundaries}
  \item{class}{Broad class of conservation practice}
  \item{cpsnum}{USDA-NRCS Conservation Practice Standard number}
  \item{cps_name}{USDA-NRCS Conservation Practice Standard name}
  \item{planner_implementation}{The implementation scenario modeled for each conservation practice standard; implementation scenarios are described in more detail in the one-page summaries associated with each practice}
  \\item{current_land_use}{Value extracted from planner_implementation to use as filtering tag.}
  \\item{irrigation}{Value extracted from planner_implementation to use as filtering tag.}
  \\item{nutrient_practice}{Value extracted from planner_implementation to use as filtering tag.}
  \item{n}{Sample size for COMET-Farm modeled practices}
  \item{lrr_fill}{Indicator column to denote when a value is derived from the LRR average. Value of 1 indicates an LRR average}
  \item{co2_mean}{Mean total carbon dioxide (CO2) emission reductions (sum of all CO2 sources) in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{co2_sterr}{Standard error of the mean of total carbon dioxide (CO2) emission reductions (sum of all CO2 sources) in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{n2o_mean}{Mean total nitrous oxide (N2O) emission reductions (sum of all N2O emission sources) in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{n2o_sterr}{Standard error of the mean for total nitrous oxide (N2O) emission reductions (sum of all N2O emission sources) in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{soil_carbon_co2}{Mean carbon dioxide (CO2) emissions reductions from soil carbon in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{soil_carbon_co2_sterr}{Standard error of the mean for carbon dioxide (CO2) emissions reductions from soil carbon in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{total_ghg_co2}{Mean of total greenhouse gas emission reductions in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  \item{total_ghg_co2_sterr}{Standard error of the mean for total greenhouse gas emission reductions in metric tonnes CO2 equivalent per acre per year (MT CO2e/ac/yr)}
  ...
}
}
\source{
\url{http://www.comet-planner.com/}
}
\usage{
comet_wa
}
\description{
A dataset containing the coefficients of greenhouse gas emissions
and soil carbon for Washington State.
}
\keyword{datasets}
